+++ modules/tinymce/src/core/test/ts/webdriver/keyboard/SpaceKeyTest.ts
@@ -1,9 +1,11 @@
-import { RealKeys } from '@ephox/agar';
+import { ApproxStructure, RealKeys } from '@ephox/agar';
 import { beforeEach, context, describe, it } from '@ephox/bedrock-client';
 import { PlatformDetection } from '@ephox/sand';
 import { TinyAssertions, TinyHooks, TinySelections } from '@ephox/wrap-mcagar';
 
 import Editor from 'tinymce/core/api/Editor';
+import { applyCaretFormat } from 'tinymce/core/fmt/CaretFormat';
+import * as Zwsp from 'tinymce/core/text/Zwsp';
 
 describe('webdriver.tinymce.core.keyboard.SpaceKeyTest', () => {
   const hook = TinyHooks.bddSetupLight<Editor>({
@@ -90,6 +92,48 @@ describe('webdriver.tinymce.core.keyboard.SpaceKeyTest', () => {
       TinySelections.setCursor(editor, [ 0, 1 ], 0);
       await RealKeys.pSendKeysOn('iframe => body', [ RealKeys.text(' ') ]);
       TinyAssertions.assertContent(editor, '<p><strong><span style="display: block;" contenteditable="false">a</span></strong>&nbsp;s</p>');
+    });
+
+    it('TINY-9310: deleting a formatted part and then press space and then insert a char should insert the char in the correct place', async () => {
+      const editor = hook.editor();
+      editor.setContent('<p>a</p>');
+      TinySelections.setCursor(editor, [ 0 ], 1);
+      applyCaretFormat(editor, 'bold', {});
+      await RealKeys.pSendKeysOn('iframe => body', [ RealKeys.text('b') ]);
+      await RealKeys.pSendKeysOn('iframe => body', [ RealKeys.backspace() ]);
+
+      TinyAssertions.assertContentStructure(editor, ApproxStructure.build((s, str) => {
+        return s.element('body', {
+          children: [
+            s.element('p', {
+              children: [
+                s.text(str.is('a')),
+                s.element('span', {
+                  attrs: {
+                    'id': str.is('_mce_caret'),
+                    'data-mce-bogus': str.is('1')
+                  },
+                  children: [
+                    s.element('strong', {
+                      children: [
+                        s.text(str.is(Zwsp.ZWSP))
+                      ]
+                    })
+                  ]
+                })
+              ]
+            })
+          ]
+        });
+      }));
+
+      await RealKeys.pSendKeysOn('iframe => body', [ RealKeys.text(' ') ]);
+      await RealKeys.pSendKeysOn('iframe => body', [ RealKeys.text('c') ]);
+      if (isFirefox) {
+        TinyAssertions.assertContent(editor, '<p>a<strong> c</strong></p>');
+      } else {
+        TinyAssertions.assertContent(editor, '<p>a c</p>');
+      }
     });
   });
 });