+++ modules/tinymce/src/plugins/accordion/main/ts/Plugin.ts
@@ -1,11 +1,13 @@
 import PluginManager from 'tinymce/core/api/PluginManager';
 
 import * as Commands from './api/Commands';
+import * as Keyboard from './core/Keyboard';
 import * as Buttons from './ui/Buttons';
 
 export default (): void => {
   PluginManager.add('accordion', (editor) => {
     Buttons.register(editor);
     Commands.register(editor);
+    Keyboard.setup(editor);
   });
 };